sriov-provider,csrcreator: use golang ephemeral container
Sriov lane jobs installes Go every time a job runs [1] [2]
in order to build and run csrcreator, which cinteracts with
the cluster api to request a certificate for sriov-opetor
webhooks.

Using ephemeral Go container could reduce the time it takes
for download and installing Go every time a job starts.
Also it will make it easier to manage the version we use.

[1] https://github.com/kubevirt/project-infra/blob/master/github/ci/prow/files/jobs/kubevirt/kubevirtci/kubevirtci-presubmits.yaml#L80
[2] https://github.com/kubevirt/project-infra/blob/master/github/ci/prow/files/jobs/kubevirt/kubevirt/kubevirt-presubmits.yaml#L207

Signed-off-by: Or Mergi <ormergi@redhat.com>