sriov provider, cert generator: use golang ephemeral container
Sriov lane jobs installes Go every time a job runs [1] [2]
in order to build and run csrcreator, which cinteracts with
the cluster api to request a certificate for sriov-opetor
webhooks.

Using ephemeral Go container could reduce the time it takes
for download and installing Go every time a job starts.
Also it will make it easier to manage the version we use.

This commit uses Go inside a container 'golang:1.13.8-alpine3.11'
image is the smallest image.
Since we certificate genrator code includes interacting with the cluster
API it is necessary to change kubeconfig to have the right IP address
of the custer API.

[1] https://github.com/kubevirt/project-infra/blob/master/github/ci
[2] https://github.com/kubevirt/project-infra/blob/master/github/ci

Signed-off-by: Or Mergi <ormergi@redhat.com>